+++ src/main/java/command/ListCommand.java
@@ -1,11 +1,16 @@
-package bug;
+package command;
+
+import exception.BugException;
+import storage.Storage;
+import task.TaskList;
+import ui.Ui;
 
 /**
  * Represents the command to list all tasks in the task management application.
  * This command displays the tasks from the current task list if there are any tasks.
  * If the task list is empty, an exception is thrown.
  */
-public class ListCommand extends Command{
+public class ListCommand extends Command {
 
     /**
      * Executes the ListCommand by displaying all tasks in the task list.
