+++ src/main/java/task/Deadlines.java
@@ -1,4 +1,4 @@
-package bug;
+package task;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
