+++ src/main/java/task/Events.java
@@ -1,4 +1,4 @@
-package bug;
+package task;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
