+++ src/main/java/task/TaskList.java
@@ -1,4 +1,4 @@
-package bug;
+package task;
 
 import java.util.List;
 import java.util.ArrayList;
