+++ src/test/java/bug/StorageTest.java
@@ -3,6 +3,8 @@ package bug;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import storage.Storage;
+import task.*;
 
 import java.io.FileWriter;
 import java.io.IOException;
